Fix several points in season preview
Send access token to use logged in state
Fix loading animation on add buttons
Fix fetch method that normally limits to 50
Add adult content filter
Temporarily fix v-menu of sorting

The temporary fix is due to Vuetify having a bug
in their v-menu script that causes activators
with tooltips to let the menu appear in
a corner instead of the mouse position.
Until that bug is fixed in a future vuetify
version, this temporary fix will stay.

Related: #148 #151 #152 #153 #154